refactor(EmailComposerInput): move to typescript
